Migrate most migrations to TS
